refactor: change files directory/imports
