+++ src/components/ui/DatePicker/DatePicker.tsx
@@ -1,9 +1,8 @@
 import { Calendar } from './Calendar/Calendar'
 import { twMerge } from 'tailwind-merge'
 import { useEffect, useRef, useState } from 'react'
-import { CalendarDay, getDaysForCalendar } from './utils/getDaysForCalendar'
-import { useCalendarSelection } from './utils/useCalendarSelection'
-import { getCalendarPositionStyles } from './utils/getCalendarPosition'
+import { CalendarDay, getCalendarPositionStyles, getDaysForCalendar } from '@/utils'
+import { useCalendarSelection } from '@/hooks/useCalendarSelection'
 import { ButtonTrigger } from './ButtonTrigger'
 import { format } from 'date-fns'
 import { normalizeDateToMidnightUTC } from '@/lib/feature/profile/ui/components/profile/GeneralInformation/GeneralInformation'
