--- src/components/ui/DatePicker/utils/getCalendarPosition.ts
@@ -1,6 +1,0 @@
-export const getCalendarPositionStyles = (rect: DOMRect | undefined) => {
-  return {
-    top: `top-[${Math.floor(rect?.top + 39)}px]`,
-    left: `left-[${Math.floor(rect?.left)}px]`,
-  }
-}
